+++ lib/src/core/buffer/buffer.dart
@@ -18,10 +18,15 @@ class Buffer {
 
   final bool isAltBuffer;
 
+  /// Characters that break selection when calling [getWordBoundary]. If null,
+  /// defaults to [defaultWordSeparators].
+  final Set<int>? wordSeparators;
+
   Buffer(
     this.terminal, {
     required this.maxLines,
     required this.isAltBuffer,
+    this.wordSeparators,
   }) {
     for (int i = 0; i < terminal.viewHeight; i++) {
       lines.push(_newEmptyLine());
@@ -464,7 +469,7 @@ class Buffer {
     return line;
   }
 
-  static final _kWordSeparators = <int>{
+  static final defaultWordSeparators = <int>{
     0,
     r' '.codeUnitAt(0),
     r'.'.codeUnitAt(0),
@@ -479,6 +484,7 @@ class Buffer {
   };
 
   BufferRangeLine? getWordBoundary(CellOffset position) {
+    var separators = wordSeparators ?? defaultWordSeparators;
     if (position.y >= lines.length) {
       return null;
     }
@@ -492,7 +498,7 @@ class Buffer {
         break;
       }
       final char = line.getCodePoint(start - 1);
-      if (_kWordSeparators.contains(char)) {
+      if (separators.contains(char)) {
         break;
       }
       start--;
@@ -503,7 +509,7 @@ class Buffer {
         break;
       }
       final char = line.getCodePoint(end);
-      if (_kWordSeparators.contains(char)) {
+      if (separators.contains(char)) {
         break;
       }
       end++;